factor out top-level fk code into `fk` package
* move all of `fk` code out of `fluidkeys/*.go` (`master` package) and
into `./fk/*.go`
* make a new `cmd/fk/main.go` which *only* passes through to `fk.Main()`
* simplify `Makefile`: build now requires single file `cmd/fk/main.go`
